+++ regionmask/core/mask.py
@@ -75,10 +75,6 @@ def _mask(
     elif method == "shapely":
         mask = _mask_shapely(lon, lat, outlines, numbers=numbers)
 
-    if np.all(np.isnan(mask)):
-        msg = "All elements of mask are NaN. Try to set 'wrap_lon=True'."
-        print(msg)
-
     if xarray is None:
         xarray = True
     else:
@@ -96,6 +92,100 @@ def _mask(
             mask = _create_xarray_2D(mask, lon_or_obj, lat_orig, lon_name, lat_name)
 
     return mask
+
+
+def _mask_2D(
+    outlines,
+    regions_is_180,
+    numbers,
+    lon_or_obj,
+    lat=None,
+    lon_name="lon",
+    lat_name="lat",
+    method=None,
+    xarray=None,
+    wrap_lon=None,
+):
+
+    mask = _mask(
+        outlines=outlines,
+        regions_is_180=regions_is_180,
+        numbers=numbers,
+        lon_or_obj=lon_or_obj,
+        lat=lat,
+        lon_name=lon_name,
+        lat_name=lat_name,
+        method=method,
+        xarray=xarray,
+        wrap_lon=wrap_lon,
+    )
+
+    if np.all(np.isnan(mask)):
+        msg = "No gridpoint belongs to any region. Returning an all-NaN mask."
+        warnings.warn(msg, UserWarning, stacklevel=3)
+
+    return mask
+
+
+def _mask_3D(
+    outlines,
+    regions_is_180,
+    numbers,
+    lon_or_obj,
+    lat=None,
+    drop=True,
+    lon_name="lon",
+    lat_name="lat",
+    method=None,
+    wrap_lon=None,
+):
+
+    mask = _mask(
+        outlines=outlines,
+        regions_is_180=regions_is_180,
+        numbers=numbers,
+        lon_or_obj=lon_or_obj,
+        lat=lat,
+        lon_name=lon_name,
+        lat_name=lat_name,
+        method=method,
+        wrap_lon=wrap_lon,
+    )
+
+    isnan = np.isnan(mask.values)
+
+    if drop:
+        numbers = np.unique(mask.values[~isnan])
+        numbers = numbers.astype(np.int)
+
+    # if no regions are found return a 0 x lat x lon mask
+    if len(numbers) == 0:
+        mask = mask.expand_dims("region", axis=0).sel(region=slice(0, 0))
+        msg = (
+            "No gridpoint belongs to any region. Returning an empty mask"
+            " with shape {}".format(mask.shape)
+        )
+        warnings.warn(msg, UserWarning, stacklevel=3)
+        return mask
+
+    mask_3D = list()
+    for num in numbers:
+        mask_3D.append(mask == num)
+
+    from distutils.version import LooseVersion
+
+    # "override" is faster but was only introduced in version 0.13.0 of xarray
+    compat = "override" if LooseVersion(xr.__version__) >= "0.13.0" else "equals"
+
+    mask_3D = xr.concat(mask_3D, dim="region", compat=compat, coords="minimal")
+
+    mask_3D = mask_3D.assign_coords(region=("region", numbers))
+
+    if np.all(isnan):
+        msg = "No gridpoint belongs to any region. Returning an all-False mask."
+        warnings.warn(msg, UserWarning, stacklevel=3)
+
+    return mask_3D
 
 
 def _determine_method(lon, lat):
@@ -186,7 +276,7 @@ def create_mask_contains(lon, lat, coords, fill=np.NaN, numbers=None):
 
     msg = (
         "The function `create_mask_contains` is deprecated and will be removed in a"
-        "  future version. Please use ``regionmask.Regions(coords).mask(lon, lat)``"
+        " future version. Please use ``regionmask.Regions(coords).mask(lon, lat)``"
         " instead."
     )
     warnings.warn(msg, FutureWarning, stacklevel=3)
@@ -256,10 +346,11 @@ def _parse_input(lon, lat, coords, fill, numbers):
     if numbers is None:
         numbers = range(n_coords)
     else:
-        assert len(numbers) == n_coords
+        if len(numbers) != n_coords:
+            raise ValueError("`numbers` and `coords` must have the same length")
 
-    msg = "The fill value should not be one of the region numbers."
-    assert fill not in numbers, msg
+    if fill in numbers:
+        raise ValueError("The fill value should not be one of the region numbers.")
 
     return lon, lat, numbers