+++ regionmask/core/regions.py
@@ -12,7 +12,7 @@ import six
 from shapely.geometry import MultiPolygon, Polygon
 
 from .formatting import _display
-from .mask import _mask
+from .mask import _mask_2D, _mask_3D
 from .plot import _plot, _plot_regions
 from .utils import _is_180, _is_numeric, _maybe_to_dict, _sanitize_names_abbrevs
 
@@ -263,7 +263,7 @@ class Regions(object):
         Parameters
         ----------
         max_rows : int, optional
-            Maximum number of rows to display in the console. Note that this add_geometries
+            Maximum number of rows to display in the console. Note that this does
             not affect the displayed metadata.
         max_width : int, optional
             Width to wrap a line in characters. If none uses console width.
@@ -311,6 +311,7 @@ class Regions(object):
             Name of latgitude in 'lon_or_obj'. Default: 'lat'
         method : None | "rasterize" | "shapely" | "legacy"
             Set method used to determine wether a gridpoint lies in a region.
+            If None (default) autoselects the method depending on the grid spacing.
         xarray : None | bool, optional
             Deprecated. If None or True returns an xarray DataArray, if False returns a
             numpy ndarray. Default: None.
@@ -339,7 +340,7 @@ class Regions(object):
         else:
             outlines = self.polygons
 
-        return _mask(
+        return _mask_2D(
             outlines=outlines,
             regions_is_180=self.lon_180,
             numbers=self.numbers,
@@ -351,6 +352,82 @@ class Regions(object):
             xarray=xarray,
             wrap_lon=wrap_lon,
         )
+
+    def mask_3D(
+        self,
+        lon_or_obj,
+        lat=None,
+        drop=True,
+        lon_name="lon",
+        lat_name="lat",
+        method=None,
+        wrap_lon=None,
+    ):
+        """
+        create a 3D boolean mask of a set of regions for the given lat/ lon grid
+
+        Parameters
+        ----------
+        lon_or_obj : object or array_like
+            Can either be a longitude array and then ``lat`` needs to be
+            given. Or an object where the longitude and latitude can be
+            retrived as: ``lon = lon_or_obj[lon_name]`` and
+            ``lat = lon_or_obj[lat_name]``
+        lat : array_like, optional
+            If ``lon_or_obj`` is a longitude array, the latitude needs to be
+            specified here.
+        drop : boolean, optional
+            If True (default) drops slices where all elements are False (i.e no
+            gridpoints are contained in a region). If False returns one slice per
+            region.
+        lon_name : str, optional
+            Name of longitude in 'lon_or_obj'. Default: "lon".
+        lat_name : str, optional
+            Name of latgitude in 'lon_or_obj'. Default: "lat"
+        method : None | "rasterize" | "shapely", optional
+            Set method used to determine wether a gridpoint lies in a region.
+            If None (default) autoselects the method depending on the grid spacing.
+        wrap_lon : None | bool | 180 | 360, optional
+            Whether to wrap the longitude around, should be inferred automatically.
+            If the regions and the provided longitude do not have the same
+            base (i.e. one is -180..180 and the other 0..360) one of them
+            must be wrapped. This can be done with wrap_lon.
+            If wrap_lon is None autodetects whether the longitude needs to be
+            wrapped. If wrap_lon is False, nothing is done. If wrap_lon is True,
+            longitude data is wrapped to 360 if its minimum is smaller
+            than 0 and wrapped to 180 if its maximum is larger than 180.
+
+        Returns
+        -------
+        mask_3D : boolean xarray.DataArray
+
+        """
+
+        if method == "legacy":
+            raise ValueError("method 'legacy' not supported in 'mask_3D'")
+
+        mask_3D = _mask_3D(
+            outlines=self.polygons,
+            regions_is_180=self.lon_180,
+            numbers=self.numbers,
+            lon_or_obj=lon_or_obj,
+            lat=lat,
+            drop=drop,
+            lon_name=lon_name,
+            lat_name=lat_name,
+            method=method,
+            wrap_lon=wrap_lon,
+        )
+
+        numbers = mask_3D.region.values
+        abbrevs = self[numbers].abbrevs
+        names = self[numbers].names
+
+        mask_3D = mask_3D.assign_coords(
+            abbrevs=("region", abbrevs), names=("region", names)
+        )
+
+        return mask_3D
 
 
 # add the plotting methods
@@ -497,8 +574,11 @@ class _OneRegion(object):
             self._polygon = None
             outline = np.asarray(outline)
 
-            assert outline.ndim == 2, "Outline must be 2D"
-            assert outline.shape[1] == 2, "Outline must have Nx2 elements"
+            if outline.ndim != 2:
+                raise ValueError("Outline must be 2D")
+
+            if outline.shape[1] != 2:
+                raise ValueError("Outline must have Nx2 elements")
 
             self._coords = np.array(outline)